feat: melhorias no codigo
