+++ package.json
@@ -28,6 +28,7 @@
     "react-icons": "^5.4.0",
     "react-native": "^0.77.0",
     "react-router": "^7.0.2",
+    "react-router-dom": "^7.1.3",
     "styled-components": "^6.1.13",
     "use-mask-input": "^3.4.2"
   },
@@ -42,7 +43,8 @@
     "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.14",
-    "@types/react": "^18.3.12",
+    "@types/node": "^22.10.10",
+    "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.1",
     "@typescript-eslint/eslint-plugin": "^8.17.0",
     "@typescript-eslint/parser": "^8.17.0",
