+++ src/pages/Loans/LoansHeader/header.tsx
@@ -1,0 +1,7 @@
+const Header = () => (
+    <div id='header' style={{ backgroundColor: '#D9D9D9', height: '15vh', width: '100vw', display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
+        <img src="public/logo.png" alt="Logo" style={{ height: '7vh' }} />
+    </div>
+);
+
+export default Header;
